feat(switch): re-enable all 12 deferred files; full Switch build 100% green
Wire every previously-deferred TU back into the Switch CMake manifest and
resolve the remaining hex-rays 64-bit / strict-aliasing / SIMD-mask issues:

* EffectsCore: fx_beam, fx_convert, fx_draw, fx_marks, fx_system, fx_update,
  fx_update_util — all landed
* xanim: dobj, xanim — landed (dobj save/restore capped at sizeof(SavedDObj)
  to avoid OOB on 64-bit; KISAKHACK-AUDIT marker preserves runtime concern)
* ui_shared_obj — landed (Eval expression evaluator stack laundered via
  KisakEval_GetDouble/SetDouble/OpAt noinline helpers; opStack[1024] +
  valStack[1024] adjacent indexing past [1024] expressed through OpAt)
* cgame_mp/cg_ents_mp — landed (SIMD AABB rotation laundered via
  Kisak_PackFloat2/UnpackFloatLo/Hi memcpy helpers)
* bgame/bg_animation_mp — landed

Supporting changes:
* Add _QWORD_alias may-alias typedef in q_shared.h next to existing
  _BYTE/_WORD/_DWORD aliases; bulk-replace  casts across the
  affected files so GCC strict-aliasing accepts the upstream byte semantics
* Fix LONG typedef: was  (8 bytes on aarch64), now  to match
  Win32 LONG width and prevent 8-byte atomic ops on 4-byte fields
* Cascade fix:  →  in r_rendercmds.h,
  fxprimitives.h, r_scene.h, xanim.h, fx_system/update/sort.cpp
* Cascade fix:  in Interlocked* call sites →
  static_cast<LONG>(...) across 10 files
* fx_marks v26 scratch: _DWORD[60] → float[60] (it actually holds float
  matrix/quaternion data, hex-rays mistyped)
* fx_beam swizzle: Kisak_Float4ByteAt helper for byte-swizzle past PackedUnitVec.array[4]
* Various sign-compare / array-bounds / class-memaccess fixes in 30+ files
* Final NRO artifact still produces (5.8 MB), POSIX build remains green

Note: Many of the SIMD-style rewrites preserve the exact byte semantics the
upstream relied on (via memcpy/may-alias). The runtime behavior should match
the upstream 32-bit binary on byte arithmetic, but the larger struct sizes
(pointer widening) shift some nega-array reaches off their original targets.
KISAKHACK-AUDIT markers point at every such site.

+++ src/gfx_d3d/r_drawsurf.cpp
@@ -45,7 +45,7 @@ char __cdecl R_ReserveCodeMeshArgs(int argCount, unsigned int* argOffsetOut)
     if ((unsigned int)(argCount + oldArgCount) < 0x100)
     {
         *argOffsetOut = oldArgCount;
-        InterlockedExchange(&frontEndDataOut->codeMeshArgsCount, static_cast<long>(argCount + oldArgCount));
+        InterlockedExchange(&frontEndDataOut->codeMeshArgsCount, static_cast<LONG>(argCount + oldArgCount));
         return 1;
     }
     else
@@ -100,7 +100,7 @@ void __cdecl R_AddCodeMeshDrawSurf(
 
     if (Material_GetTechnique(material, gfxDrawMethod.emissiveTechType) && (material->info.gameFlags & 2) == 0)
     {
-        codeMeshIndex = InterlockedExchangeAdd(&frontEndDataOut->codeMeshCount, static_cast<long>(1));
+        codeMeshIndex = InterlockedExchangeAdd(&frontEndDataOut->codeMeshCount, static_cast<LONG>(1));
         if (codeMeshIndex < 0x800)
         {
             localCodeMesh = &frontEndDataOut->codeMeshes[codeMeshIndex];
@@ -236,7 +236,7 @@ void __cdecl R_AddMarkMeshDrawSurf(
     iassert(rgp.sortedMaterials[material->info.drawSurf.fields.materialSortedIndex] == material);
     if (Material_GetTechnique(material, (MaterialTechniqueType)gfxDrawMethod.litTechType[11][0]))
     {
-        markMeshIndex = InterlockedExchangeAdd(&frontEndDataOut->markMeshCount, static_cast<long>(1));
+        markMeshIndex = InterlockedExchangeAdd(&frontEndDataOut->markMeshCount, static_cast<LONG>(1));
         if (markMeshIndex < 0x600)
         {
             markMesh = &frontEndDataOut->markMeshes[markMeshIndex];
@@ -533,7 +533,7 @@ char __cdecl R_AddParticleCloudDrawSurf(unsigned int cloudIndex, Material *mater
     [[maybe_unused]] GfxDrawSurf *drawSurf; // [esp+2Ch] [ebp-Ch]
     [[maybe_unused]] int region; // [esp+30h] [ebp-8h]
 
-    if (cloudIndex >= frontEndDataOut->cloudCount)
+    if (cloudIndex >= static_cast<unsigned int>(frontEndDataOut->cloudCount))
         MyAssertHandler(
             ".\\r_drawsurf.cpp",
             561,
